% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_polygon_membership}
\alias{add_polygon_membership}
\title{Add Polygon Membership to Seurat Object}
\usage{
add_polygon_membership(seurat_obj, polygons, fov_name = "fov")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{polygons}{sf polygon object}

\item{fov_name}{FOV name}
}
\value{
Seurat object with added metadata
}
\description{
Adds metadata column indicating if cells are inside polygons
}
