% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{launch_stGradient}
\alias{launch_stGradient}
\title{Launch stGradient Interactive Polygon Selector}
\usage{
launch_stGradient()
}
\value{
Launches Shiny application
}
\description{
Opens a Shiny app for interactively drawing polygons on spatial data
}
\examples{
\dontrun{
launch_stGradient()
}
}
