% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{launch_tunes}
\alias{launch_tunes}
\title{Launch TuNeS Interactive Polygon Selector}
\usage{
launch_tunes()
}
\value{
Launches Shiny application
}
\description{
Opens a Shiny app for interactively drawing polygons on spatial data
}
\examples{
\dontrun{
launch_tunes()
}
}
