% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_svg_path}
\alias{parse_svg_path}
\title{Parse SVG Path to Coordinates}
\usage{
parse_svg_path(path_string)
}
\arguments{
\item{path_string}{SVG path string from plotly}
}
\value{
Matrix of x,y coordinates
}
\description{
Internal function to parse plotly SVG paths
}
\keyword{internal}
