% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_all_metrics}
\alias{plot_all_metrics}
\title{Plot All Metrics Faceted}
\usage{
plot_all_metrics(distance_profile_combined)
}
\arguments{
\item{distance_profile_combined}{Combined data frame from both modes}
}
\value{
ggplot object
}
\description{
Creates faceted plot of all separation metrics
}
