% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_celltype_heatmap}
\alias{plot_celltype_heatmap}
\title{Plot Cell Type Contributions Heatmap}
\usage{
plot_celltype_heatmap(celltype_contributions)
}
\arguments{
\item{celltype_contributions}{Data frame from calculate_distance_profile}
}
\value{
ggplot object
}
\description{
Shows which cell types drive differences at each distance
}
