% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_celltype_proportions}
\alias{plot_celltype_proportions}
\title{Plot Cell Type Proportions}
\usage{
plot_celltype_proportions(celltype_contributions, top_n = 5)
}
\arguments{
\item{celltype_contributions}{Data frame from calculate_distance_profile}

\item{top_n}{Number of top cell types to show}
}
\value{
ggplot object
}
\description{
Shows proportion trajectories for top cell types
}
