% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_distance_comparison}
\alias{plot_distance_comparison}
\title{Plot Distance Profile Comparison}
\usage{
plot_distance_comparison(distance_profile_combined)
}
\arguments{
\item{distance_profile_combined}{Combined data frame from both modes}
}
\value{
ggplot object
}
\description{
Plots transcriptomic distance for different inside modes
}
