% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_analysis.R
\name{plot_niche_coefficients}
\alias{plot_niche_coefficients}
\title{Plot Niche Logistic Coefficients}
\usage{
plot_niche_coefficients(univar_res, title = "Niche logistic coefficients")
}
\arguments{
\item{univar_res}{Output from `run_niche_univariate_lr`.}

\item{title}{Plot title.}
}
\value{
ggplot object.
}
\description{
Plot Niche Logistic Coefficients
}
