% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_analysis.R
\name{plot_niche_stacked_composition}
\alias{plot_niche_stacked_composition}
\title{Plot Patient Niche Stacked Composition}
\usage{
plot_niche_stacked_composition(niche_prop_df)
}
\arguments{
\item{niche_prop_df}{Output `niche_prop_df` from `run_niche_pipeline`.}
}
\value{
ggplot object.
}
\description{
Plot Patient Niche Stacked Composition
}
