% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_analysis.R
\name{plot_pcf_outside_heatmap}
\alias{plot_pcf_outside_heatmap}
\title{Plot Outside-Polygon PCF Heatmap}
\usage{
plot_pcf_outside_heatmap(pcf_summary, midpoint = 1)
}
\arguments{
\item{pcf_summary}{Output from `calculate_pcf_outside_summary`.}

\item{midpoint}{Midpoint value for diverging color scale.}
}
\value{
ggplot object.
}
\description{
Converts PCF summary output into a symmetric matrix and visualizes it as a heatmap.
}
